A bunch of shit

+++ server/src/main.ts
@@ -2,6 +2,8 @@ import * as express from 'express';
 import * as socket_io from 'socket.io';
 import * as file_upload from 'express-fileupload';
 import * as fs from 'fs';
+import { stringify } from 'querystring';
+import { userInfo } from 'os';
 
 let app = express();
 let http = require('http').Server(app);
@@ -32,13 +34,68 @@ class User {
 };
 
 class Round {
+	setup: string;
+	puns: Array<{user_id: string, pun: string}>;
+	all_puns_submitted: boolean;
+	all_votes_submitted: boolean;
+	winners_id: Array<string>;
+	votes: Array<{user_id: string, voted_user_id: string}>; // user ids
 
+	constructor() {
+		this.puns = new Array<{user_id: string, pun: string}>();
+		this.all_puns_submitted = false;
+		this.votes = new Array<{user_id: string, voted_user_id: string}>(); 
+		this.winners_id = new Array<string>();
+	}
+
+	addPun = (user_id: string, pun: string) : void => {
+		this.puns.push({user_id, pun});
+	}
+
+	hasUserPunned = (user_id: string) : boolean => {
+		let result = this.puns.filter((pun_user) => {
+			return pun_user.user_id === user_id;
+		});
+		return result.length > 0;
+	}
+
+	vote = (user_id: string, vote_id: string) : void => {
+		this.votes.push({user_id: user_id, voted_user_id: vote_id});
+	}
+
+	calcWinner = () : Array<string> => {
+		let largest_vote_count = 0;
+		for(let i = 0; i < this.puns.length; i++) {
+			let user_id = this.puns[i].user_id;
+			let result = this.votes.filter((vote) => {
+				return vote.voted_user_id == user_id;
+			});
+			if(result.length > largest_vote_count) {
+				largest_vote_count = result.length;
+			}
+		}
+
+		let results = new Array<string>();
+
+		for(let i = 0; i < this.puns.length; i++) {
+			let user_id = this.puns[i].user_id;
+			let result = this.votes.filter((vote) => {
+				return vote.voted_user_id == user_id;
+			});
+			if(result.length === largest_vote_count) {
+				results.push(user_id);
+			}
+		}
+
+		return results;
+		
+	}
 };
 
 class Room {
 	id: number;
 	users: Array<User>;
-	current_rount: Round;
+	current_round: Round;
 	admin_id: string;
 
     constructor(id: number) {
@@ -48,7 +105,7 @@ class Room {
 	
 	startRound = () : void => {
 		let round: Round = new Round();
-		this.current_rount = round;
+		this.current_round = round;
 		this.admin_id = this.users[Math.floor(Math.random() * this.users.length)].id;
 		
 		this.update();
@@ -63,14 +120,19 @@ class Room {
     }
 
     removeUser = (user: User) => {
-		fs.unlink(user.image_path, () => {console.log("User photo deleted")});
-		let index = this.users.indexOf(user);
+		if(user !== undefined) {
+			fs.unlink(user.image_path, () => {console.log("User photo deleted")});
+			let index = this.users.indexOf(user);
 
-		if(index > -1) {
-			this.users.splice(index, 1);
+			if(index > -1) {
+				this.users.splice(index, 1);
+			} else {
+				
+			}
 		} else {
-			
+			console.log("Attempting to delete undefined user");
 		}
+		
 		
 		console.log(this.users);
 		this.update();
@@ -96,7 +158,33 @@ class Room {
 	}
 
 	isJoinable = () : boolean => {
-		return this.current_rount == undefined || this.current_rount == null;
+		return this.current_round == undefined || this.current_round == null;
+	}
+
+	allPunsSubmitted = () : boolean => {
+		let result = true;
+		for(let i = 0; i < this.users.length; i++) {
+			if(this.users[i].id == this.admin_id) continue;
+			if(!this.current_round.hasUserPunned(this.users[i].id)) {
+				result = false;
+				break;
+			}
+		}
+		return result;
+	}
+
+	allVotesSubmitted = () : boolean => {
+		let result = true;
+		for(let i = 0; i < this.users.length; i++) {
+			let results: Array<{user_id: string, voted_user_id: string}> = this.current_round.votes.filter((vote) => {
+				return vote.user_id == this.users[i].id;
+			});
+			if(results.length === 0) {
+				result = false;
+				break;
+			}
+		}
+		return result;
 	}
 };
 
@@ -114,23 +202,29 @@ io.on("connection", (socket) => {
 	let user: User = undefined;
 	let room: Room = undefined;
     socket.on("join_room", (data, callback) => {
-        room = getRoomFromID(data);
+		console.log("Trying to find room");
+		room = getRoomFromID(data);
+		console.log("Found room");
         if(room !== undefined && room.isJoinable()) {
+			console.log("Joining room");
             socket.join(data);
-            user = room.addUser(socket);
+			user = room.addUser(socket);
+			console.log("joined room");
             callback(true, user.id);
         } else {
             callback(false);
-        }
+		}
+		console.log(room);
         
 	});
 	
-	socket.on("view_room", (room_id) => {
+	socket.on("view_room", (room_id, callback) => {
+		console.log("Viewing room");
 		socket.join(room_id);
 		let room: Room = getRoomFromID(room_id);
 		if(room !== undefined) {
 			console.log("Emitting room update");
-			room.update();
+			callback(room);
 		}
 	});
 
@@ -138,7 +232,29 @@ io.on("connection", (socket) => {
 		let room: Room = getRoomFromID(room_id);
 		room.startRound();
 	})
+
+	socket.on("set_round_setup", (setup) => {
+		room.current_round.setup = setup;
+		room.update();
+		room.emitEvent("players_begin");
+	})
+
+	socket.on("pun_submit", (pun) => {
+		room.current_round.addPun(user.id, pun);
+		if(room.allPunsSubmitted()) {
+			room.current_round.all_puns_submitted = true;
+		}
+		room.update();
+	})
 	
+	socket.on("vote", (user_id) => {
+		room.current_round.vote(user.id, user_id);
+		if(room.allVotesSubmitted()) {
+			room.current_round.all_votes_submitted = true;
+		}
+		room.update();
+	})
+
 	socket.on("disconnect", () => {
 		console.log("User disconnected");
 		if(room !== undefined)
